ValidateExecutableVisitor use a single, shared, introspection query
Doing the introspection query all the time is hurting performance,
this does not change, so a single query can be done with all the
fields, build a validation tree and then query it on all validations.

This is the first part, moving the top-level query to be read-only.

The next commits will eliminate other `executeQuery()`, then a shared
context will be created and hosted by the `Response` class, which can
be discovered using introspection or fed using JSON (schema.json).