refactor ValidateExecutableVisitor, split field get x add
Split the fields getter and cache/insertion into the map so they can
be used later in a different way.

There should be no code differences, just moving around the internal
branch to its own function.

In the next commits, this will be removed from the getter, as it will
be query-only as the types will be all cached beforehand.