add specific ValidateType
Instead of using a map with properties `name`, `kind` (string) and
`ofType` (another map), use a set of custom classes with kind as
enumeration and ofType as shared pointer.

This allows much simpler handling and comparison, no need to serialize
to string to make it simpler to compare.

We can also store reference to types, know which kind (ie:
isInputType?) and save memory by using references, in particular to
common types such as Int, Float, String...

The matchingTypes and fields are stored as part of each ValidateType
specialization.