simplify variable processing
Instead of 2 maps + set (both ordered), use one single unordered_map
(string_view) + unordered_set (pointer to definition).

The string_view is okay since the ast_node tree is valid during the
processing, so the references are valid.

The pointer to definition is also okay for _referencedVariables, since
the defintitions are all created upfront and the map (thus references)
won't change while visiting the fields.

The 2->1 map was possible since we're now storing the definition
location instead of using a second map just to store the ast_node to
query the position on errors.