optimize Request::findOperationDefinition()
We don't need to validate this again, since the validation will
already cover the specs bits. This way we save doing the whole loop,
allocating usedNames and the queries that were already performed
during validation.

In our internal usages, there is no need to allocate the string, keep
only the string_view to the operationType.

Keep the old function, but mark as deprected.