Validation: replace queue with list
Our pattern to quickly push and pop some items to be processed is not
that good with std::queue, which uses deque (double-ended queue). This
is because it will keep an internal array to allow O(1) access.

Since we do not use direct field access, let's stick with a
double-linked list