Prepare Request to receive a ValidationContext
This is basically moving code around, change the parameters to allow
the request to receive the ValidationContext.

Bring back the original Request() constructor so it will not break
dynamically linked binaries.

The introspection results are kept around, in the future the
validation context will only use pointers to strings (string_view) and
everything must be alive in order to work.