BREAKING SelectionSetParams keeps a reference to the parent
Change SelectionSetParams to keep an optional reference to the parent,
this way we don't need to build the path over and over again just to
add one element, instead we create this on demand.

Since errorPath was accessed directly, this breaks the existing code,
it became a method that dynamically computes the error path
(recursive).

This is important since we just pay the list copy price when there is
an error and not in all field resolution.