BREAKING SelectionSetParams optimized
SelectionSetParams now references the parent, errorPath changed.

Instead of copying the path over and over again for each
SelectionSetParams, keep a reference to the parent and a
`ownErrorPath` with the current segment, then compute the full path on
demand. This breaks the structure by keeping the parent reference and
`ownErrorPath`, `errorPath()` is now a getter/accessor.

ResolverParams fieldName is now a view: Since we keep a reference to
`field` (ast_node) is a child node, either the `peg::field_name` or
`peg::alias_name` string views, then there is no reason we should be
copying this.

Last but not least, avoid the extra copy to create ResolverParams and
delete SelectionSetParams() copy constructor