Object::resolve() create a single visitor
There is no reason we should be doing a new SelectionVisitor per
field, this even helps with the values, as we can store them in the
same location.

Even the names (set), they can't be repeated in the same object
selection.

Also allows us to pass the expected number of fields (reservedValues),
this is a bit wrong since the fragments may contain more than one
child, but overall this should provide a good estimate.